backend/requirements.txt
flask==3.1.0
flask-socketio==5.5.1
flask-cors==5.0.1
python-chess==1.999
eventlet==0.37.0
